remove base_tx.go
